Cast fix

+++ libsrc/leddevice/dev_serial/LedDeviceKarate.cpp
@@ -46,7 +46,7 @@ bool LedDeviceKarate::init(const QJsonObject &deviceConfig)
 
 int LedDeviceKarate::write(const std::vector<ColorRgb> &ledValues)
 {
-	for (signed iLed=0; iLed<_ledCount; iLed++)
+	for (signed iLed=0; iLed< static_cast<int>(_ledCount); iLed++)
         {
 		
        		const ColorRgb& rgb = ledValues[iLed];
